requirements.txt
asgiref==3.8.1
Django==5.2
django-cors-headers==4.4.0
djangorestframework==3.15.1
djangorestframework-simplejwt==5.5.0
PyJWT==2.9.0
sqlparse==0.5.1
typing-extensions==4.12.2
gunicorn==22.0.0